Se agrega tabla para usuarios facebook. Se crean los login con google y facebook usando oauth 2.0 authorization code grant y openid connect

+++ database/migrations/20200615015615-create-google-user.js
@@ -6,11 +6,11 @@ module.exports = {
         allowNull: false,
         autoIncrement: true,
         primaryKey: true,
-        type: Sequelize.INTEGER
+        type: Sequelize.BIGINT
       },
       googleId: {
         allowNull: false,
-        type: Sequelize.INTEGER
+        type: Sequelize.BIGINT
       },
       name: {
         type: Sequelize.STRING,
